+++ src/comutl/AXP_Utility.h
@@ -41,11 +41,11 @@
 
 typedef unsigned char		u8;
 typedef unsigned short		u16;
-typedef unsigned long		u32;
+typedef unsigned int		u32;
 typedef unsigned long long	u64;
 typedef char				i8;
 typedef short				i16;
-typedef long				i32;
+typedef int					i32;
 typedef long long			i64;
 
 typedef struct
